extract tracking functions
